Move component test to unit
